Refactor hooks and business logic
