+++ app/src/components/ConfirmParticipant/ConfirmParticipant.tsx
@@ -1,5 +1,4 @@
 import React from 'react';
-import { useEvent } from 'src/hooks/eventHooks';
 import { stringifyDate } from 'src/types/date';
 import { stringifyTime } from 'src/types/time';
 import style from './ConfirmParticipant.module.scss';
@@ -7,6 +6,7 @@ import { viewEventRoute, eventIdKey, emailKey } from 'src/routing';
 import { hasLoaded } from 'src/remote-data';
 import { BlockLink } from 'src/components/Common/BlockLink/BlockLink';
 import { useParam } from 'src/utils/browser-state';
+import { useEvent } from 'src/hooks/cache';
 import { Page } from 'src/components/Page/Page';
 
 export const ConfirmParticipant = () => {
